+++ src/Svg.Skia/Svg/GroupDrawable.cs
@@ -1,0 +1,66 @@
+﻿#if !USE_PICTURE
+using SkiaSharp;
+using CropRect = SkiaSharp.SKImageFilter.CropRect;
+#endif
+#if USE_PICTURE
+using SKRect = Svg.Picture.Rect;
+#endif
+
+namespace Svg.Skia
+{
+    internal sealed class GroupDrawable : DrawableContainer
+    {
+        private GroupDrawable()
+            : base()
+        {
+        }
+
+        public static GroupDrawable Create(SvgGroup svgGroup, SKRect skOwnerBounds, DrawableBase? parent, Attributes ignoreAttributes = Attributes.None)
+        {
+            var drawable = new GroupDrawable
+            {
+                Element = svgGroup,
+                Parent = parent,
+                IgnoreAttributes = ignoreAttributes
+            };
+
+            drawable.IsDrawable = drawable.CanDraw(svgGroup, drawable.IgnoreAttributes) && drawable.HasFeatures(svgGroup, drawable.IgnoreAttributes);
+
+            // NOTE: Call AddMarkers only once.
+            SvgExtensions.AddMarkers(svgGroup);
+
+            drawable.CreateChildren(svgGroup, skOwnerBounds, drawable, ignoreAttributes);
+
+            // TODO: Check if children are explicitly set to be visible.
+            //foreach (var child in drawable.ChildrenDrawables)
+            //{
+            //    if (child.IsDrawable)
+            //    {
+            //        IsDrawable = true;
+            //        break;
+            //    }
+            //}
+
+            if (!drawable.IsDrawable)
+            {
+                return drawable;
+            }
+
+            drawable.IsAntialias = SvgExtensions.IsAntialias(svgGroup);
+
+            drawable.TransformedBounds = SKRect.Empty;
+
+            drawable.CreateTransformedBounds();
+
+            drawable.Transform = SvgExtensions.ToSKMatrix(svgGroup.Transforms);
+
+            drawable.Fill = null;
+            drawable.Stroke = null;
+
+            // TODO: Transform _skBounds using _skMatrix.
+            drawable.TransformedBounds = drawable.Transform.MapRect(drawable.TransformedBounds);
+
+            return drawable;
+        }
+    }
+}
